Feature: file upload button
Drag and drop is not always the most convenient way to upload a file.
Sometimes old-fashioned open dialog is preferred. Moreover, open file
dialog allows to filter file type, this it will be easier for user to
understand which file they should select.

In this commit I add file upload button to /upload endpoint, so user
could either drag-n-drop files or choose them old-fashioned way.
